+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - feat: Add `SetRequest` method on a `Scope` to control `Request` context data
 - feat: Add `FromHTTPRequest` for `Request` type for easier extraction
 - ref: Extract `Request` information more accurately
+- fix: Attach `ServerName`, `Release`, `Dist`, `Environment` options to the event
 - fix: Don't log events dropped due to full transport buffer as sent
 - fix: Don't panic and create an appropriate event when called `CaptureException` or `Recover` with `nil` value